계산기 [STEP2] 찰스

+++ Calculator/Calculator/Models/CalculateItem.swift
@@ -9,6 +9,6 @@ protocol CalculateItem {
     
 }
 
-extension Int: CalculateItem {
+extension Double: CalculateItem {
     
 }
